requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==5.0.1
cachetools==5.3.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.6
Django==4.1.5
django-allauth==0.52.0
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-post-office==3.6.3
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
facebook-sdk==3.1.0
google==3.0.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
kombu==5.2.4
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.4.0
priority==1.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.1.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.12.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tinycss2==1.1.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
zope.interface==5.5.2
